updated for MR
