+++ src/pages/page404/index.jsx
@@ -3,9 +3,6 @@ import Navigation from "../../components/navigation";
 export default function Page404() {
   return (
     <>
-    {/* <nav>
-      About
-    </nav> */}
     <Navigation />
     <main>
       <h2>
